analysis-server/requirements.txt
amqp==2.6.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.18.40
botocore==1.21.40
celery==4.4.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dateparser==1.0.0
demjson==2.2.4
Django==3.2.5
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
flower==0.9.7
humanize==3.11.0
jmespath==0.10.0
kombu==4.6.11
maya==0.6.1
pendulum==2.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.20
pycurl==7.43.0.5
PyMySQL==1.0.2
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
regex==2021.8.28
s3transfer==0.5.0
six==1.16.0
snaptime==0.2.4
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
tzlocal==3.0
urllib3==1.26.6
vine==1.3.0
wcwidth==0.2.5
